BaseOptions definition moved to CLI module
