+++ src/cli/__tests__/check.test.ts
@@ -1,5 +1,5 @@
-import { BaseOptions } from "../../cli"
-import { check } from "../check"
+import { BaseOptions } from '../options'
+import { check } from '../check'
 
 const options: BaseOptions = {
   averageDays: true,
@@ -11,20 +11,20 @@ const options: BaseOptions = {
   percentage: false,
   silent: false,
   totalDays: false,
-  totalPackages: false,
+  totalPackages: false
 }
 
-describe("CLI args check", () => {
-  test("throws an error when maxDate value is not a valid date", () => {
+describe('CLI args check', () => {
+  test('throws an error when maxDate value is not a valid date', () => {
     expect(() => {
-      check({...options, maxDate: 'abc'})
+      check({ ...options, maxDate: 'abc' })
     }).toThrow(/date 'abc' is not valid/)
   })
 
   test.each([
     options,
-    {...options, maxDate: '2000-01-19'}
-  ])("returns true when all arguments have valid values #%#", opts => {
+    { ...options, maxDate: '2000-01-19' }
+  ])('returns true when all arguments have valid values #%#', opts => {
     expect(check(opts)).toBe(true)
   })
 })
