+++ src/cli/options.ts
@@ -1,0 +1,12 @@
+export interface BaseOptions {
+  averageDays: boolean
+  devOnly: boolean
+  dirtyPackages: boolean
+  excludeDev: boolean
+  format: string
+  maxDate: string | null
+  percentage: boolean
+  silent: boolean
+  totalDays: boolean
+  totalPackages: boolean
+}
